Refactor presenters
