+++ src/presenter/header-presenter.js
@@ -1,6 +1,6 @@
 import {render} from '../framework/render.js';
 import TripInfoView from '../view/trip-info-view.js';
-import TripFiltersView, {DEFAULT_FILTER_NAME} from '../view/trip-filters-view.js';
+import FiltersView, {DEFAULT_FILTER_NAME} from '../view/filters-view.js';
 import NewEventButtonView from '../view/new-event-button-view.js';
 import {generateFilters} from '../utils.js';
 
@@ -18,7 +18,7 @@ export default class HeaderPresenter {
 
   init() {
     this.#renderTripInfo();
-    this.#renderTripFilters();
+    this.#renderFilters();
     this.#renderNewEventButton();
   }
 
@@ -28,9 +28,9 @@ export default class HeaderPresenter {
     }
   }
 
-  #renderTripFilters() {
+  #renderFilters() {
     const filters = generateFilters(this.#events, this.#activeFilterName);
-    render(new TripFiltersView(filters), this.headerContainer);
+    render(new FiltersView(filters), this.headerContainer);
   }
 
   #renderNewEventButton() {
